+++ R/tar_dir.R
@@ -1,7 +1,8 @@
 #' @title Execute code in a temporary directory.
 #' @export
-#' @family utilities to extend targets
-#' @description Runs code inside a new `tempfile()` directory
+#' @keywords internal
+#' @description Not a user-side function. Just for CRAN.
+#' @details Runs code inside a new `tempfile()` directory
 #'   in order to avoid writing to the user's file space.
 #'   Used in examples and tests in order to comply with CRAN policies.
 #' @return Return value of the user-defined code.