+++ R/tar_package.R
@@ -35,7 +35,7 @@
 #'   packageVersion stack tail
 #' @importFrom vctrs vec_c vec_rbind vec_size vec_slice
 #' @importFrom withr local_dir local_envvar
-#'   local_options with_dir with_seed
+#'   local_options with_seed
 #' @importFrom yaml read_yaml
 NULL